+++ test/bump.test.ts
@@ -1,6 +1,6 @@
 // tslint:disable: max-line-length
-import * as cdk from "monocdk";
-import { AutoBump, WritableGitHubRepo } from "../lib";
+import * as cdk from 'monocdk';
+import { AutoBump, WritableGitHubRepo } from '../lib';
 import '@monocdk-experiment/assert/jest';
 
 const Stack = cdk.Stack;
@@ -10,7 +10,7 @@ const MOCK_REPO = new WritableGitHubRepo({
   commitUsername: 'user',
   commitEmail: 'email@email',
   repository: 'owner/repo',
-  tokenSecretArn: 'token-secret-arn'
+  tokenSecretArn: 'token-secret-arn',
 });
 
 test('autoBump', () => {
@@ -19,7 +19,7 @@ test('autoBump', () => {
 
   // WHEN
   new AutoBump(stack, 'MyAutoBump', {
-    repo: MOCK_REPO
+    repo: MOCK_REPO,
   });
 
   // THEN
@@ -27,43 +27,43 @@ test('autoBump', () => {
   // build project
   expect(stack).toHaveResource('AWS::CodeBuild::Project', {
     Triggers: {
-      Webhook: false
+      Webhook: false,
     },
     Source: {
       Type: 'GITHUB',
       GitCloneDepth: 0,
-      Location: "https://github.com/owner/repo.git",
+      Location: 'https://github.com/owner/repo.git',
       ReportBuildStatus: false,
       BuildSpec: JSON.stringify({
-        "version": "0.2",
-        "phases": {
-          "pre_build": {
-            "commands": [
-              "git config --global user.email \"email@email\"",
-              "git config --global user.name \"user\""
-            ]
+        version: '0.2',
+        phases: {
+          pre_build: {
+            commands: [
+              'git config --global user.email "email@email"',
+              'git config --global user.name "user"',
+            ],
           },
-          "build": {
-            "commands": [
-              "export SKIP=false",
-              "$SKIP || { aws secretsmanager get-secret-value --secret-id \"ssh-key-secret-arn\" --output=text --query=SecretString > ~/.ssh/id_rsa ; }",
-              "$SKIP || { mkdir -p ~/.ssh ; }",
-              "$SKIP || { chmod 0600 ~/.ssh/id_rsa ~/.ssh/config ; }",
-              "$SKIP || { ssh-keyscan -t rsa github.com >> ~/.ssh/known_hosts ; }",
-              "$SKIP || { ls .git && { echo \".git directory exists\";  } || { echo \".git directory doesnot exist - cloning...\" && git init . && git remote add origin git@github.com:owner/repo.git && git fetch && git reset --hard origin/master && git branch -M master && git clean -fqdx; } ; }",
+          build: {
+            commands: [
+              'export SKIP=false',
+              '$SKIP || { aws secretsmanager get-secret-value --secret-id "ssh-key-secret-arn" --output=text --query=SecretString > ~/.ssh/id_rsa ; }',
+              '$SKIP || { mkdir -p ~/.ssh ; }',
+              '$SKIP || { chmod 0600 ~/.ssh/id_rsa ~/.ssh/config ; }',
+              '$SKIP || { ssh-keyscan -t rsa github.com >> ~/.ssh/known_hosts ; }',
+              '$SKIP || { ls .git && { echo ".git directory exists";  } || { echo ".git directory doesnot exist - cloning..." && git init . && git remote add origin git@github.com:owner/repo.git && git fetch && git reset --hard origin/master && git branch -M master && git clean -fqdx; } ; }',
               "$SKIP || { git describe --exact-match master && { echo 'Skip condition is met, skipping...' && export SKIP=true; } || { echo 'Skip condition is not met, continuing...' && export SKIP=false; } ; }",
-              "$SKIP || { git rev-parse --verify origin/bump/$VERSION && { git checkout bump/$VERSION && git merge master && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands;  } || { git checkout master && git checkout -b temp && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands && git branch -M bump/$VERSION; } ; }",
-              "$SKIP || { git merge-base --is-ancestor bump/$VERSION origin/master && { echo \"Skipping: bump/$VERSION is an ancestor of origin/master\"; export SKIP=true; } || { echo \"Pushing: bump/$VERSION is ahead of origin/master\"; export SKIP=false; } ; }",
-              "$SKIP || { git remote add origin_ssh git@github.com:owner/repo.git ; }",
-              "$SKIP || { git push --follow-tags origin_ssh bump/$VERSION:bump/$VERSION ; }",
-              "$SKIP || { export GITHUB_TOKEN=$(aws secretsmanager get-secret-value --secret-id \"token-secret-arn\" --output=text --query=SecretString) ; }",
+              '$SKIP || { git rev-parse --verify origin/bump/$VERSION && { git checkout bump/$VERSION && git merge master && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands;  } || { git checkout master && git checkout -b temp && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands && git branch -M bump/$VERSION; } ; }',
+              '$SKIP || { git merge-base --is-ancestor bump/$VERSION origin/master && { echo "Skipping: bump/$VERSION is an ancestor of origin/master"; export SKIP=true; } || { echo "Pushing: bump/$VERSION is ahead of origin/master"; export SKIP=false; } ; }',
+              '$SKIP || { git remote add origin_ssh git@github.com:owner/repo.git ; }',
+              '$SKIP || { git push --follow-tags origin_ssh bump/$VERSION:bump/$VERSION ; }',
+              '$SKIP || { export GITHUB_TOKEN=$(aws secretsmanager get-secret-value --secret-id "token-secret-arn" --output=text --query=SecretString) ; }',
               "$SKIP || { curl --fail -X POST -o pr.json --header \"Authorization: token $GITHUB_TOKEN\" --header \"Content-Type: application/json\" -d \"{\\\"title\\\":\\\"chore(release): $VERSION\\\",\\\"base\\\":\\\"master\\\",\\\"head\\\":\\\"bump/$VERSION\\\"}\" https://api.github.com/repos/owner/repo/pulls && export PR_NUMBER=$(node -p 'require(\"./pr.json\").number') ; }",
-              "$SKIP || { curl --fail -X PATCH --header \"Authorization: token $GITHUB_TOKEN\" --header \"Content-Type: application/json\" -d \"{\\\"body\\\":\\\"See [CHANGELOG](https://github.com/owner/repo/blob/bump/$VERSION/CHANGELOG.md)\\\"}\" https://api.github.com/repos/owner/repo/pulls/$PR_NUMBER ; }"
-            ]
-          }
-        }
-      }, undefined, 2)
-    }
+              '$SKIP || { curl --fail -X PATCH --header "Authorization: token $GITHUB_TOKEN" --header "Content-Type: application/json" -d "{\\"body\\":\\"See [CHANGELOG](https://github.com/owner/repo/blob/bump/$VERSION/CHANGELOG.md)\\"}" https://api.github.com/repos/owner/repo/pulls/$PR_NUMBER ; }',
+            ],
+          },
+        },
+      }, undefined, 2),
+    },
   });
 
 });
@@ -75,12 +75,12 @@ test('autoBump with schedule', () => {
   // WHEN
   new AutoBump(stack, 'MyAutoBump', {
     repo: MOCK_REPO,
-    scheduleExpression: "cron(0 12 * * ? *)"
+    scheduleExpression: 'cron(0 12 * * ? *)',
   });
 
   // default schedule
   expect(stack).toHaveResource('AWS::Events::Rule', {
-    ScheduleExpression: "cron(0 12 * * ? *)"
+    ScheduleExpression: 'cron(0 12 * * ? *)',
   });
 
 });
@@ -92,7 +92,7 @@ test('autoBump with custom cloneDepth', () => {
   // WHEN
   new AutoBump(stack, 'MyAutoBump', {
     repo: MOCK_REPO,
-    cloneDepth: 10
+    cloneDepth: 10,
   });
 
   // THEN
@@ -100,43 +100,43 @@ test('autoBump with custom cloneDepth', () => {
   // build project
   expect(stack).toHaveResource('AWS::CodeBuild::Project', {
     Triggers: {
-      Webhook: false
+      Webhook: false,
     },
     Source: {
       Type: 'GITHUB',
       GitCloneDepth: 10,
-      Location: "https://github.com/owner/repo.git",
+      Location: 'https://github.com/owner/repo.git',
       ReportBuildStatus: false,
       BuildSpec: JSON.stringify({
-        "version": "0.2",
-        "phases": {
-          "pre_build": {
-            "commands": [
-              "git config --global user.email \"email@email\"",
-              "git config --global user.name \"user\""
-            ]
+        version: '0.2',
+        phases: {
+          pre_build: {
+            commands: [
+              'git config --global user.email "email@email"',
+              'git config --global user.name "user"',
+            ],
           },
-          "build": {
-            "commands": [
-              "export SKIP=false",
-              "$SKIP || { aws secretsmanager get-secret-value --secret-id \"ssh-key-secret-arn\" --output=text --query=SecretString > ~/.ssh/id_rsa ; }",
-              "$SKIP || { mkdir -p ~/.ssh ; }",
-              "$SKIP || { chmod 0600 ~/.ssh/id_rsa ~/.ssh/config ; }",
-              "$SKIP || { ssh-keyscan -t rsa github.com >> ~/.ssh/known_hosts ; }",
-              "$SKIP || { ls .git && { echo \".git directory exists\";  } || { echo \".git directory doesnot exist - cloning...\" && git init . && git remote add origin git@github.com:owner/repo.git && git fetch && git reset --hard origin/master && git branch -M master && git clean -fqdx; } ; }",
+          build: {
+            commands: [
+              'export SKIP=false',
+              '$SKIP || { aws secretsmanager get-secret-value --secret-id "ssh-key-secret-arn" --output=text --query=SecretString > ~/.ssh/id_rsa ; }',
+              '$SKIP || { mkdir -p ~/.ssh ; }',
+              '$SKIP || { chmod 0600 ~/.ssh/id_rsa ~/.ssh/config ; }',
+              '$SKIP || { ssh-keyscan -t rsa github.com >> ~/.ssh/known_hosts ; }',
+              '$SKIP || { ls .git && { echo ".git directory exists";  } || { echo ".git directory doesnot exist - cloning..." && git init . && git remote add origin git@github.com:owner/repo.git && git fetch && git reset --hard origin/master && git branch -M master && git clean -fqdx; } ; }',
               "$SKIP || { git describe --exact-match master && { echo 'Skip condition is met, skipping...' && export SKIP=true; } || { echo 'Skip condition is not met, continuing...' && export SKIP=false; } ; }",
-              "$SKIP || { git rev-parse --verify origin/bump/$VERSION && { git checkout bump/$VERSION && git merge master && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands;  } || { git checkout master && git checkout -b temp && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands && git branch -M bump/$VERSION; } ; }",
-              "$SKIP || { git merge-base --is-ancestor bump/$VERSION origin/master && { echo \"Skipping: bump/$VERSION is an ancestor of origin/master\"; export SKIP=true; } || { echo \"Pushing: bump/$VERSION is ahead of origin/master\"; export SKIP=false; } ; }",
-              "$SKIP || { git remote add origin_ssh git@github.com:owner/repo.git ; }",
-              "$SKIP || { git push --follow-tags origin_ssh bump/$VERSION:bump/$VERSION ; }",
-              "$SKIP || { export GITHUB_TOKEN=$(aws secretsmanager get-secret-value --secret-id \"token-secret-arn\" --output=text --query=SecretString) ; }",
+              '$SKIP || { git rev-parse --verify origin/bump/$VERSION && { git checkout bump/$VERSION && git merge master && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands;  } || { git checkout master && git checkout -b temp && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands && git branch -M bump/$VERSION; } ; }',
+              '$SKIP || { git merge-base --is-ancestor bump/$VERSION origin/master && { echo "Skipping: bump/$VERSION is an ancestor of origin/master"; export SKIP=true; } || { echo "Pushing: bump/$VERSION is ahead of origin/master"; export SKIP=false; } ; }',
+              '$SKIP || { git remote add origin_ssh git@github.com:owner/repo.git ; }',
+              '$SKIP || { git push --follow-tags origin_ssh bump/$VERSION:bump/$VERSION ; }',
+              '$SKIP || { export GITHUB_TOKEN=$(aws secretsmanager get-secret-value --secret-id "token-secret-arn" --output=text --query=SecretString) ; }',
               "$SKIP || { curl --fail -X POST -o pr.json --header \"Authorization: token $GITHUB_TOKEN\" --header \"Content-Type: application/json\" -d \"{\\\"title\\\":\\\"chore(release): $VERSION\\\",\\\"base\\\":\\\"master\\\",\\\"head\\\":\\\"bump/$VERSION\\\"}\" https://api.github.com/repos/owner/repo/pulls && export PR_NUMBER=$(node -p 'require(\"./pr.json\").number') ; }",
-              "$SKIP || { curl --fail -X PATCH --header \"Authorization: token $GITHUB_TOKEN\" --header \"Content-Type: application/json\" -d \"{\\\"body\\\":\\\"See [CHANGELOG](https://github.com/owner/repo/blob/bump/$VERSION/CHANGELOG.md)\\\"}\" https://api.github.com/repos/owner/repo/pulls/$PR_NUMBER ; }"
-            ]
-          }
-        }
-      }, undefined, 2)
-    }
+              '$SKIP || { curl --fail -X PATCH --header "Authorization: token $GITHUB_TOKEN" --header "Content-Type: application/json" -d "{\\"body\\":\\"See [CHANGELOG](https://github.com/owner/repo/blob/bump/$VERSION/CHANGELOG.md)\\"}" https://api.github.com/repos/owner/repo/pulls/$PR_NUMBER ; }',
+            ],
+          },
+        },
+      }, undefined, 2),
+    },
   });
 });
 
@@ -147,12 +147,12 @@ test('autoBump with schedule disabled', () => {
   // WHEN
   new AutoBump(stack, 'MyAutoBump', {
     repo: MOCK_REPO,
-    scheduleExpression: 'disable'
+    scheduleExpression: 'disable',
   });
 
   // THEN
   expect(stack).not.toHaveResource('AWS::Events::Rule', {
-    ScheduleExpression: "cron(0 12 * * ? *)"
+    ScheduleExpression: 'cron(0 12 * * ? *)',
   });
 });
 
@@ -164,13 +164,13 @@ test('autoBump with push only', () => {
     commitUsername: 'user',
     commitEmail: 'email@email',
     repository: 'owner/repo',
-    tokenSecretArn: 'token-secret-arn'
+    tokenSecretArn: 'token-secret-arn',
   });
 
   // WHEN
   new AutoBump(stack, 'MyAutoBump', {
     repo,
-    pushOnly: true
+    pushOnly: true,
   });
 
   // THEN
@@ -178,40 +178,40 @@ test('autoBump with push only', () => {
   // build project
   expect(stack).toHaveResource('AWS::CodeBuild::Project', {
     Triggers: {
-      Webhook: false
+      Webhook: false,
     },
     Source: {
       Type: 'GITHUB',
       GitCloneDepth: 0,
-      Location: "https://github.com/owner/repo.git",
+      Location: 'https://github.com/owner/repo.git',
       ReportBuildStatus: false,
       BuildSpec: JSON.stringify({
-        "version": "0.2",
-        "phases": {
-          "pre_build": {
-            "commands": [
-              "git config --global user.email \"email@email\"",
-              "git config --global user.name \"user\""
-            ]
+        version: '0.2',
+        phases: {
+          pre_build: {
+            commands: [
+              'git config --global user.email "email@email"',
+              'git config --global user.name "user"',
+            ],
           },
-          "build": {
-            "commands": [
-              "export SKIP=false",
-              "$SKIP || { aws secretsmanager get-secret-value --secret-id \"ssh-key-secret-arn\" --output=text --query=SecretString > ~/.ssh/id_rsa ; }",
-              "$SKIP || { mkdir -p ~/.ssh ; }",
-              "$SKIP || { chmod 0600 ~/.ssh/id_rsa ~/.ssh/config ; }",
-              "$SKIP || { ssh-keyscan -t rsa github.com >> ~/.ssh/known_hosts ; }",
-              "$SKIP || { ls .git && { echo \".git directory exists\";  } || { echo \".git directory doesnot exist - cloning...\" && git init . && git remote add origin git@github.com:owner/repo.git && git fetch && git reset --hard origin/master && git branch -M master && git clean -fqdx; } ; }",
+          build: {
+            commands: [
+              'export SKIP=false',
+              '$SKIP || { aws secretsmanager get-secret-value --secret-id "ssh-key-secret-arn" --output=text --query=SecretString > ~/.ssh/id_rsa ; }',
+              '$SKIP || { mkdir -p ~/.ssh ; }',
+              '$SKIP || { chmod 0600 ~/.ssh/id_rsa ~/.ssh/config ; }',
+              '$SKIP || { ssh-keyscan -t rsa github.com >> ~/.ssh/known_hosts ; }',
+              '$SKIP || { ls .git && { echo ".git directory exists";  } || { echo ".git directory doesnot exist - cloning..." && git init . && git remote add origin git@github.com:owner/repo.git && git fetch && git reset --hard origin/master && git branch -M master && git clean -fqdx; } ; }',
               "$SKIP || { git describe --exact-match master && { echo 'Skip condition is met, skipping...' && export SKIP=true; } || { echo 'Skip condition is not met, continuing...' && export SKIP=false; } ; }",
-              "$SKIP || { git rev-parse --verify origin/bump/$VERSION && { git checkout bump/$VERSION && git merge master && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands;  } || { git checkout master && git checkout -b temp && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands && git branch -M bump/$VERSION; } ; }",
-              "$SKIP || { git merge-base --is-ancestor bump/$VERSION origin/master && { echo \"Skipping: bump/$VERSION is an ancestor of origin/master\"; export SKIP=true; } || { echo \"Pushing: bump/$VERSION is ahead of origin/master\"; export SKIP=false; } ; }",
-              "$SKIP || { git remote add origin_ssh git@github.com:owner/repo.git ; }",
-              "$SKIP || { git push --follow-tags origin_ssh bump/$VERSION:bump/$VERSION ; }"
-            ]
-          }
-        }
-      }, undefined, 2)
-    }
+              '$SKIP || { git rev-parse --verify origin/bump/$VERSION && { git checkout bump/$VERSION && git merge master && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands;  } || { git checkout master && git checkout -b temp && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands && git branch -M bump/$VERSION; } ; }',
+              '$SKIP || { git merge-base --is-ancestor bump/$VERSION origin/master && { echo "Skipping: bump/$VERSION is an ancestor of origin/master"; export SKIP=true; } || { echo "Pushing: bump/$VERSION is ahead of origin/master"; export SKIP=false; } ; }',
+              '$SKIP || { git remote add origin_ssh git@github.com:owner/repo.git ; }',
+              '$SKIP || { git push --follow-tags origin_ssh bump/$VERSION:bump/$VERSION ; }',
+            ],
+          },
+        },
+      }, undefined, 2),
+    },
   });
 });
 
@@ -226,8 +226,8 @@ test('autoBump with pull request with custom options', () => {
     title: 'custom title',
     body: 'custom body',
     base: {
-      name: 'release'
-    }
+      name: 'release',
+    },
 
   });
 
@@ -236,43 +236,43 @@ test('autoBump with pull request with custom options', () => {
   // build project
   expect(stack).toHaveResource('AWS::CodeBuild::Project', {
     Triggers: {
-      Webhook: false
+      Webhook: false,
     },
     Source: {
       Type: 'GITHUB',
       GitCloneDepth: 0,
-      Location: "https://github.com/owner/repo.git",
+      Location: 'https://github.com/owner/repo.git',
       ReportBuildStatus: false,
       BuildSpec: JSON.stringify({
-        "version": "0.2",
-        "phases": {
-          "pre_build": {
-            "commands": [
-              "git config --global user.email \"email@email\"",
-              "git config --global user.name \"user\""
-            ]
+        version: '0.2',
+        phases: {
+          pre_build: {
+            commands: [
+              'git config --global user.email "email@email"',
+              'git config --global user.name "user"',
+            ],
           },
-          "build": {
-            "commands": [
-              "export SKIP=false",
-              "$SKIP || { aws secretsmanager get-secret-value --secret-id \"ssh-key-secret-arn\" --output=text --query=SecretString > ~/.ssh/id_rsa ; }",
-              "$SKIP || { mkdir -p ~/.ssh ; }",
-              "$SKIP || { chmod 0600 ~/.ssh/id_rsa ~/.ssh/config ; }",
-              "$SKIP || { ssh-keyscan -t rsa github.com >> ~/.ssh/known_hosts ; }",
-              "$SKIP || { ls .git && { echo \".git directory exists\";  } || { echo \".git directory doesnot exist - cloning...\" && git init . && git remote add origin git@github.com:owner/repo.git && git fetch && git reset --hard origin/release && git branch -M release && git clean -fqdx; } ; }",
+          build: {
+            commands: [
+              'export SKIP=false',
+              '$SKIP || { aws secretsmanager get-secret-value --secret-id "ssh-key-secret-arn" --output=text --query=SecretString > ~/.ssh/id_rsa ; }',
+              '$SKIP || { mkdir -p ~/.ssh ; }',
+              '$SKIP || { chmod 0600 ~/.ssh/id_rsa ~/.ssh/config ; }',
+              '$SKIP || { ssh-keyscan -t rsa github.com >> ~/.ssh/known_hosts ; }',
+              '$SKIP || { ls .git && { echo ".git directory exists";  } || { echo ".git directory doesnot exist - cloning..." && git init . && git remote add origin git@github.com:owner/repo.git && git fetch && git reset --hard origin/release && git branch -M release && git clean -fqdx; } ; }',
               "$SKIP || { git describe --exact-match release && { echo 'Skip condition is met, skipping...' && export SKIP=true; } || { echo 'Skip condition is not met, continuing...' && export SKIP=false; } ; }",
-              "$SKIP || { git rev-parse --verify origin/bump/$VERSION && { git checkout bump/$VERSION && git merge release && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands;  } || { git checkout release && git checkout -b temp && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands && git branch -M bump/$VERSION; } ; }",
-              "$SKIP || { git merge-base --is-ancestor bump/$VERSION origin/release && { echo \"Skipping: bump/$VERSION is an ancestor of origin/release\"; export SKIP=true; } || { echo \"Pushing: bump/$VERSION is ahead of origin/release\"; export SKIP=false; } ; }",
-              "$SKIP || { git remote add origin_ssh git@github.com:owner/repo.git ; }",
-              "$SKIP || { git push --follow-tags origin_ssh bump/$VERSION:bump/$VERSION ; }",
-              "$SKIP || { export GITHUB_TOKEN=$(aws secretsmanager get-secret-value --secret-id \"token-secret-arn\" --output=text --query=SecretString) ; }",
+              '$SKIP || { git rev-parse --verify origin/bump/$VERSION && { git checkout bump/$VERSION && git merge release && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands;  } || { git checkout release && git checkout -b temp && /bin/sh ./bump.sh && export VERSION=$(git describe) && echo Finished running user commands && git branch -M bump/$VERSION; } ; }',
+              '$SKIP || { git merge-base --is-ancestor bump/$VERSION origin/release && { echo "Skipping: bump/$VERSION is an ancestor of origin/release"; export SKIP=true; } || { echo "Pushing: bump/$VERSION is ahead of origin/release"; export SKIP=false; } ; }',
+              '$SKIP || { git remote add origin_ssh git@github.com:owner/repo.git ; }',
+              '$SKIP || { git push --follow-tags origin_ssh bump/$VERSION:bump/$VERSION ; }',
+              '$SKIP || { export GITHUB_TOKEN=$(aws secretsmanager get-secret-value --secret-id "token-secret-arn" --output=text --query=SecretString) ; }',
               "$SKIP || { curl --fail -X POST -o pr.json --header \"Authorization: token $GITHUB_TOKEN\" --header \"Content-Type: application/json\" -d \"{\\\"title\\\":\\\"custom title\\\",\\\"base\\\":\\\"release\\\",\\\"head\\\":\\\"bump/$VERSION\\\"}\" https://api.github.com/repos/owner/repo/pulls && export PR_NUMBER=$(node -p 'require(\"./pr.json\").number') ; }",
-              "$SKIP || { curl --fail -X PATCH --header \"Authorization: token $GITHUB_TOKEN\" --header \"Content-Type: application/json\" -d \"{\\\"body\\\":\\\"custom body\\\"}\" https://api.github.com/repos/owner/repo/pulls/$PR_NUMBER ; }"
-            ]
-          }
-        }
-      }, undefined, 2)
-    }
+              '$SKIP || { curl --fail -X PATCH --header "Authorization: token $GITHUB_TOKEN" --header "Content-Type: application/json" -d "{\\"body\\":\\"custom body\\"}" https://api.github.com/repos/owner/repo/pulls/$PR_NUMBER ; }',
+            ],
+          },
+        },
+      }, undefined, 2),
+    },
   });
 });
 
@@ -284,10 +284,10 @@ test('autoBump with pull request fails when head=base', () => {
   expect(() => new AutoBump(stack, 'MyAutoBump', {
     repo: MOCK_REPO,
     base: {
-      name: 'master'
+      name: 'master',
     },
     head: {
-      name: 'master'
-    }
+      name: 'master',
+    },
   })).toThrow();
 });