+++ test/open-pgp-key-pair.test.ts
@@ -1,12 +1,14 @@
-import { aws_kms as kms } from "monocdk";
-import * as cdk from 'monocdk';
-import assert = require("@monocdk-experiment/assert");
-import { OpenPGPKeyPair } from "../lib/open-pgp-key-pair";
+import * as assert from '@monocdk-experiment/assert';
+import {
+  App, Stack,
+  aws_kms as kms,
+} from 'monocdk';
+import { OpenPGPKeyPair } from '../lib/open-pgp-key-pair';
 
 
 test('correctly creates', () => {
   // GIVEN
-  const stack = new cdk.Stack(new cdk.App(), 'TestStack');
+  const stack = new Stack(new App(), 'TestStack');
   const encryptionKey = new kms.Key(stack, 'CMK');
   // WHEN
   new OpenPGPKeyPair(stack, 'Secret', {
@@ -17,7 +19,7 @@ test('correctly creates', () => {
     keySizeBits: 1_024,
     pubKeyParameterName: 'TestParameter',
     secretName: 'SecretName',
-    version: 0
+    version: 0,
   });
 
   // THEN
@@ -35,7 +37,7 @@ test('correctly creates', () => {
 
 test('correctly forwards parameter name', () => {
   // GIVEN
-  const stack = new cdk.Stack(new cdk.App(), 'TestStack');
+  const stack = new Stack(new App(), 'TestStack');
   const parameterName = 'TestParameterName';
 
   // WHEN
@@ -47,7 +49,7 @@ test('correctly forwards parameter name', () => {
     identity: 'Test',
     keySizeBits: 1_024,
     secretName: 'SecretName',
-    version: 0
+    version: 0,
   });
 
   // THEN
@@ -60,7 +62,7 @@ test('correctly forwards parameter name', () => {
 
 test('Handler has appropriate permissions', () => {
   // GIVEN
-  const stack = new cdk.Stack(new cdk.App(), 'TestStack');
+  const stack = new Stack(new App(), 'TestStack');
 
   // WHEN
   new OpenPGPKeyPair(stack, 'Secret', {
@@ -88,17 +90,16 @@ test('Handler has appropriate permissions', () => {
         ],
         Resource: {
           'Fn::Join': ['',
-            ['arn:', { Ref: 'AWS::Partition' }, ':secretsmanager:', { Ref: 'AWS::Region' }, ':', { Ref: 'AWS::AccountId' }, ':secret:Bar-??????']
-          ]
-        }
+            ['arn:', { Ref: 'AWS::Partition' }, ':secretsmanager:', { Ref: 'AWS::Region' }, ':', { Ref: 'AWS::AccountId' }, ':secret:Bar-??????']],
+        },
       }, {
         Effect: 'Allow',
         Action: 'ssm:DeleteParameter',
-        Resource: '*'
-      }]
+        Resource: '*',
+      }],
     },
     PolicyName: 'SingletonLambdaf25803d3054b44fc985f4860d7d6ee74ServiceRoleDefaultPolicyA8FDF5BD',
-    Roles: [{ Ref: 'SingletonLambdaf25803d3054b44fc985f4860d7d6ee74ServiceRole410148CF' }]
+    Roles: [{ Ref: 'SingletonLambdaf25803d3054b44fc985f4860d7d6ee74ServiceRole410148CF' }],
   }));
 
   assert.expect(stack).to(assert.haveResourceLike('AWS::KMS::Key', {
@@ -111,10 +112,10 @@ test('Handler has appropriate permissions', () => {
         Action: ['kms:Decrypt', 'kms:GenerateDataKey'],
         Resource: '*',
         Condition: {
-          StringEquals: { 'kms:ViaService': { 'Fn::Join': ['', ['secretsmanager.', { Ref: 'AWS::Region' }, '.amazonaws.com']] } }
+          StringEquals: { 'kms:ViaService': { 'Fn::Join': ['', ['secretsmanager.', { Ref: 'AWS::Region' }, '.amazonaws.com']] } },
         },
         Principal: { AWS: { 'Fn::GetAtt': ['SingletonLambdaf25803d3054b44fc985f4860d7d6ee74ServiceRole410148CF', 'Arn'] } },
-      }]
-    }
+      }],
+    },
   }));
 });